requirements.txt
Django==4.1
djangorestframework
Pillow
django-ckeditor
cryptography
django-rest-knox